requirements.txt
numpy
jupyter
torch
pandas
matplotlib
seqeval
tabulate
psutil
pyyaml
nltk
beautifulsoup4
lxml
